api: add settings.go
