+++ api/bus.go
@@ -34,12 +34,6 @@ const (
 	UsabilityFilterModeUnusable = "unusable"
 
 	DefaultBucketName = "default"
-
-	SettingContractSet      = "contractset"
-	SettingGouging          = "gouging"
-	SettingRedundancy       = "redundancy"
-	SettingS3Authentication = "s3authentication"
-	SettingUploadPacking    = "uploadpacking"
 )
 
 var (
@@ -86,10 +80,6 @@ var (
 	// ErrPartNotFound is returned if the specified part of a multipart upload
 	// wasn't found.
 	ErrPartNotFound = errors.New("multipart upload part not found")
-
-	// ErrSettingNotFound is returned if a requested setting is not present in the
-	// database.
-	ErrSettingNotFound = errors.New("setting not found")
 
 	// ErrUploadAlreadyExists is returned when starting an upload with an id
 	// that's already in use.
@@ -494,54 +484,6 @@ type GougingParams struct {
 	TransactionFee     types.Currency
 }
 
-// ContractSetSetting contains the default contract set used by the worker for
-// uploads and migrations.
-type ContractSetSetting struct {
-	Default string `json:"default"`
-}
-
-// GougingSettings contain some price settings used in price gouging.
-type GougingSettings struct {
-	// MinMaxCollateral is the minimum value for 'MaxCollateral' in the host's
-	// price settings
-	MinMaxCollateral types.Currency `json:"minMaxCollateral"`
-
-	// MaxRPCPrice is the maximum allowed base price for RPCs
-	MaxRPCPrice types.Currency `json:"maxRPCPrice"`
-
-	// MaxContractPrice is the maximum allowed price to form a contract
-	MaxContractPrice types.Currency `json:"maxContractPrice"`
-
-	// MaxDownloadPrice is the maximum allowed price to download 1TiB of data
-	MaxDownloadPrice types.Currency `json:"maxDownloadPrice"`
-
-	// MaxUploadPrice is the maximum allowed price to upload 1TiB of data
-	MaxUploadPrice types.Currency `json:"maxUploadPrice"`
-
-	// MaxStoragePrice is the maximum allowed price to store 1 byte per block
-	MaxStoragePrice types.Currency `json:"maxStoragePrice"`
-
-	// HostBlockHeightLeeway is the amount of blocks of leeway given to the host
-	// block height in the host's price table
-	HostBlockHeightLeeway int `json:"hostBlockHeightLeeway"`
-
-	// MinPriceTableValidity is the minimum accepted value for `Validity` in the
-	// host's price settings.
-	MinPriceTableValidity time.Duration `json:"minPriceTableValidity"`
-
-	// MinAccountExpiry is the minimum accepted value for `AccountExpiry` in the
-	// host's price settings.
-	MinAccountExpiry time.Duration `json:"minAccountExpiry"`
-
-	// MinMaxEphemeralAccountBalance is the minimum accepted value for
-	// `MaxEphemeralAccountBalance` in the host's price settings.
-	MinMaxEphemeralAccountBalance types.Currency `json:"minMaxEphemeralAccountBalance"`
-}
-
-type S3AuthenticationSettings struct {
-	V4Keypairs map[string]string `json:"v4Keypairs"`
-}
-
 // Option types.
 type (
 	AddObjectOptions struct {
@@ -745,23 +687,6 @@ type WalletResponse struct {
 	Unconfirmed types.Currency `json:"unconfirmed"`
 }
 
-// Validate returns an error if the gouging settings are not considered valid.
-func (gs GougingSettings) Validate() error {
-	if gs.HostBlockHeightLeeway < 3 {
-		return errors.New("HostBlockHeightLeeway must be at least 3 blocks")
-	}
-	if gs.MinAccountExpiry < time.Hour {
-		return errors.New("MinAccountExpiry must be at least 1 hour")
-	}
-	if gs.MinMaxEphemeralAccountBalance.Cmp(types.Siacoins(1)) < 0 {
-		return errors.New("MinMaxEphemeralAccountBalance must be at least 1 SC")
-	}
-	if gs.MinPriceTableValidity < 10*time.Second {
-		return errors.New("MinPriceTableValidity must be at least 10 seconds")
-	}
-	return nil
-}
-
 type (
 	Bucket struct {
 		CreatedAt time.Time    `json:"createdAt"`
@@ -794,38 +719,6 @@ type SearchHostsRequest struct {
 	UsabilityMode   string            `json:"usabilityMode"`
 	AddressContains string            `json:"addressContains"`
 	KeyIn           []types.PublicKey `json:"keyIn"`
-}
-
-type UploadPackingSettings struct {
-	Enabled               bool  `json:"enabled"`
-	SlabBufferMaxSizeSoft int64 `json:"slabBufferMaxSizeSoft"`
-}
-
-// RedundancySettings contain settings that dictate an object's redundancy.
-type RedundancySettings struct {
-	MinShards   int `json:"minShards"`
-	TotalShards int `json:"totalShards"`
-}
-
-// Redundancy returns the effective storage redundancy of the
-// RedundancySettings.
-func (rs RedundancySettings) Redundancy() float64 {
-	return float64(rs.TotalShards) / float64(rs.MinShards)
-}
-
-// Validate returns an error if the redundancy settings are not considered
-// valid.
-func (rs RedundancySettings) Validate() error {
-	if rs.MinShards < 1 {
-		return errors.New("MinShards must be greater than 0")
-	}
-	if rs.TotalShards < rs.MinShards {
-		return errors.New("TotalShards must be at least MinShards")
-	}
-	if rs.TotalShards > 255 {
-		return errors.New("TotalShards must be less than 256")
-	}
-	return nil
 }
 
 type AddPartialSlabResponse struct {
